+++ scandeval/benchmarks/__init__.py
@@ -24,7 +24,6 @@ from .idt_dep import IdtDepBenchmark  # noqa
 from .fdt_pos import FdtPosBenchmark  # noqa
 from .fdt_dep import FdtDepBenchmark  # noqa
 from .mim_gold_ner import MimGoldNerBenchmark  # noqa
-from .wikiann_is import WikiannIsBenchmark  # noqa
 from .wikiann_fo import WikiannFoBenchmark  # noqa
 from .norec_is import NorecISBenchmark  # noqa
 from .norec_fo import NorecFOBenchmark  # noqa