+++ scandeval/datasets.py
@@ -636,30 +636,6 @@ def load_mim_gold_ner() -> Tuple[pd.DataFrame, pd.DataFrame,
     return X_train, X_test, y_train, y_test
 
 
-def load_wikiann_is() -> Tuple[pd.DataFrame, pd.DataFrame,
-                               pd.DataFrame, pd.DataFrame]:
-    '''Load the the Icelandic WikiANN dataset.
-
-    Returns:
-        tuple:
-            Four dataframes, `X_train`, `X_test`, `y_train` and `y_test`, where
-            `X_train` and `X_test` corresponds to the feature matrices for the
-            training and test split, respectively, and `y_train` and `y_test`
-            contains the target vectors.
-    '''
-    base_url = ('https://raw.githubusercontent.com/saattrupdan/ScandEval/'
-                'main/datasets/wikiann_is/')
-    train_url = base_url + 'train.jsonl'
-    test_url = base_url + 'test.jsonl'
-    X_train, y_train = _get_dataset_from_url(train_url,
-                                             ['doc', 'tokens'],
-                                             'ner_tags')
-    X_test, y_test = _get_dataset_from_url(test_url,
-                                           ['doc', 'tokens'],
-                                           'ner_tags')
-    return X_train, X_test, y_train, y_test
-
-
 def load_wikiann_fo() -> Tuple[pd.DataFrame, pd.DataFrame,
                                pd.DataFrame, pd.DataFrame]:
     '''Load the the Faroese WikiANN dataset.